simulated_exchange/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.0

# WebSocket and HTTP for live feeds
websocket-client>=1.6.0
requests>=2.31.0

# Configuration
pyyaml>=6.0

# Exchange APIs
ccxt>=4.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional visualization
matplotlib>=3.7.0

# Optional Parquet support
pyarrow>=12.0.0
